Milestone_4/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.24.3
